+++ release_notes.md
@@ -2,6 +2,8 @@
   - [#137](https://github.com/MehdiK/Humanizer/pull/137): Fixed grammar error in `nb-NO` resource file & added missing Norwegian resource strings (mainly `DateHumanize_*FromNow`)
   - [#135](https://github.com/MehdiK/Humanizer/pull/135): Added Swedish localization (strings, tests)
 
+- [#140](https://github.com/MehdiK/Humanizer/pull/140): Polish localization (strings, formatter, tests)
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.17.1...master)
 
 ###v1.17.1 - 2014-04-06